+++ .oxlintrc.json
@@ -20,12 +20,14 @@
       { "min": 2, "exceptions": ["i", "j", "x", "y", "z", "_", "a", "b", "t", "T"] }
     ],
     "eslint/init-declarations": "off",
+    "eslint/max-lines": ["error", { "skipComments": true, "skipBlankLines": true, "max": 300 }],
     "eslint/max-lines-per-function": [
       "error",
       { "skipComments": true, "skipBlankLines": true, "max": 150 }
     ],
     "eslint/max-params": "off",
     "eslint/max-statements": ["error", { "ignoreTopLevelFunctions": true, "max": 20 }],
+    "eslint/no-inline-comments": "off",
     "eslint/no-magic-numbers": [
       "error",
       {
@@ -75,6 +77,7 @@
     {
       "files": ["apps/evals/**/*.ts", "apps/evals/**/*.tsx"],
       "rules": {
+        "eslint/max-lines": "off",
         "eslint/no-magic-numbers": "off",
         "typescript/no-unsafe-type-assertion": "off"
       }
@@ -82,6 +85,7 @@
     {
       "files": ["**/seed/**", "**/seed.ts"],
       "rules": {
+        "eslint/max-lines": "off",
         "eslint/max-statements": "off",
         "eslint/no-magic-numbers": "off",
         "import/no-namespace": "off"
@@ -90,6 +94,7 @@
     {
       "files": ["*.test.ts"],
       "rules": {
+        "eslint/max-lines": "off",
         "eslint/max-lines-per-function": "off",
         "eslint/max-statements": "off",
         "eslint/no-magic-numbers": "off",
@@ -102,6 +107,7 @@
     {
       "files": ["packages/ui/src/components/**/*.tsx"],
       "rules": {
+        "eslint/max-lines": "off",
         "eslint/no-magic-numbers": "off",
         "import/no-namespace": "off"
       }
