+++ apps/editor/src/app/[orgSlug]/c/[lang]/[courseSlug]/_actions/content.ts
@@ -1,0 +1,49 @@
+"use server";
+
+import { updateCourse } from "@/data/courses/update-course";
+import { getErrorMessage } from "@/lib/error-messages";
+import { revalidateMainApp } from "@zoonk/core/cache/revalidate";
+import { cacheTagCourse } from "@zoonk/utils/cache";
+import { after } from "next/server";
+
+export async function updateCourseTitleAction(
+  courseSlug: string,
+  courseId: number,
+  data: { title: string },
+): Promise<{ error: string | null }> {
+  const { error } = await updateCourse({
+    courseId,
+    title: data.title,
+  });
+
+  if (error) {
+    return { error: await getErrorMessage(error) };
+  }
+
+  after(async () => {
+    await revalidateMainApp([cacheTagCourse({ courseSlug })]);
+  });
+
+  return { error: null };
+}
+
+export async function updateCourseDescriptionAction(
+  courseSlug: string,
+  courseId: number,
+  data: { description: string },
+): Promise<{ error: string | null }> {
+  const { error } = await updateCourse({
+    courseId,
+    description: data.description,
+  });
+
+  if (error) {
+    return { error: await getErrorMessage(error) };
+  }
+
+  after(async () => {
+    await revalidateMainApp([cacheTagCourse({ courseSlug })]);
+  });
+
+  return { error: null };
+}
