--- apps/editor/src/components/editor-list.tsx
@@ -1,522 +1,0 @@
-"use client";
-
-import {
-  DndContext,
-  type DragEndEvent,
-  DragOverlay,
-  type DragStartEvent,
-  KeyboardSensor,
-  PointerSensor,
-  TouchSensor,
-  closestCenter,
-  useSensor,
-  useSensors,
-} from "@dnd-kit/core";
-import {
-  SortableContext,
-  arrayMove,
-  sortableKeyboardCoordinates,
-  useSortable,
-  verticalListSortingStrategy,
-} from "@dnd-kit/sortable";
-import { CSS } from "@dnd-kit/utilities";
-import { Button } from "@zoonk/ui/components/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@zoonk/ui/components/dropdown-menu";
-import { Skeleton } from "@zoonk/ui/components/skeleton";
-import { toast } from "@zoonk/ui/components/sonner";
-import { mergeProps, useRender } from "@zoonk/ui/lib/render";
-import { cn } from "@zoonk/ui/lib/utils";
-import { isNextRedirectError } from "@zoonk/utils/error";
-import { EllipsisIcon, LoaderCircleIcon, PlusIcon } from "lucide-react";
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useId,
-  useMemo,
-  useOptimistic,
-  useState,
-  useTransition,
-} from "react";
-
-type EditorListContextValue = {
-  pending: boolean;
-  handleInsert: (position: number) => void;
-};
-
-const EditorListContext = createContext<EditorListContextValue | undefined>(undefined);
-
-function useEditorList() {
-  const context = useContext(EditorListContext);
-  if (!context) {
-    throw new Error("EditorList components must be used within an EditorListProvider.");
-  }
-  return context;
-}
-
-function EditorListProvider({
-  children,
-  onInsert,
-}: {
-  children: React.ReactNode;
-  onInsert: (position: number) => Promise<void>;
-}) {
-  const [pending, startTransition] = useTransition();
-
-  const handleInsert = useCallback(
-    (position: number) => {
-      startTransition(async () => {
-        try {
-          await onInsert(position);
-        } catch (error) {
-          if (!isNextRedirectError(error)) {
-            console.error(`Failed to insert item at position ${position}:`, error);
-
-            toast.error(error instanceof Error ? error.message : String(error));
-          }
-        }
-      });
-    },
-    [onInsert],
-  );
-
-  const value = useMemo<EditorListContextValue>(
-    () => ({
-      handleInsert,
-      pending,
-    }),
-    [handleInsert, pending],
-  );
-
-  return (
-    <EditorListContext.Provider value={value}>
-      <div className="relative" data-slot="editor-list">
-        {children}
-      </div>
-    </EditorListContext.Provider>
-  );
-}
-
-function EditorListSpinner({ className, ...props }: React.ComponentProps<"div">) {
-  const { pending } = useEditorList();
-
-  if (!pending) {
-    return null;
-  }
-
-  return (
-    <div
-      className={cn(
-        "bg-background/60 absolute inset-0 z-20 flex items-center justify-center",
-        className,
-      )}
-      data-slot="editor-list-spinner"
-      {...props}
-    >
-      <LoaderCircleIcon className="text-muted-foreground size-5 animate-spin" />
-    </div>
-  );
-}
-
-function EditorListHeader({ className, ...props }: React.ComponentProps<"div">) {
-  return (
-    <div
-      className={cn("flex items-center justify-end gap-2 px-4 pb-2", className)}
-      data-slot="editor-list-header"
-      {...props}
-    />
-  );
-}
-
-function EditorListAddButton({
-  children,
-  className,
-  position,
-  ...props
-}: Omit<React.ComponentProps<typeof Button>, "onClick"> & {
-  position: number;
-}) {
-  const { pending, handleInsert } = useEditorList();
-
-  return (
-    <Button
-      className={className}
-      data-slot="editor-list-add-button"
-      disabled={pending}
-      onClick={() => handleInsert(position)}
-      size="sm"
-      variant="outline"
-      {...props}
-    >
-      <PlusIcon />
-      {children}
-    </Button>
-  );
-}
-
-function EditorListContent({ className, ...props }: React.ComponentProps<"ul">) {
-  return <ul className={cn(className)} data-slot="editor-list-content" {...props} />;
-}
-
-function EditorListItem({ className, ...props }: React.ComponentProps<"li">) {
-  return <li className={cn(className)} data-slot="editor-list-item" {...props} />;
-}
-
-function EditorListItemLink({ className, render, ...props }: useRender.ComponentProps<"a">) {
-  return useRender({
-    defaultTagName: "a",
-    props: mergeProps<"a">(
-      {
-        className: cn("flex min-w-0 flex-1 items-start gap-4", className),
-      },
-      props,
-    ),
-    render,
-    state: {
-      slot: "editor-list-item-link",
-    },
-  });
-}
-
-function EditorListItemContent({ className, ...props }: React.ComponentProps<"div">) {
-  return (
-    <div
-      className={cn("min-w-0 flex-1", className)}
-      data-slot="editor-list-item-content"
-      {...props}
-    />
-  );
-}
-
-function EditorListItemTitle({ className, ...props }: React.ComponentProps<"p">) {
-  return (
-    <p
-      className={cn("leading-tight font-medium", className)}
-      data-slot="editor-list-item-title"
-      {...props}
-    />
-  );
-}
-
-function EditorListItemDescription({ className, ...props }: React.ComponentProps<"p">) {
-  return (
-    <p
-      className={cn("text-muted-foreground mt-0.5 line-clamp-2 text-sm", className)}
-      data-slot="editor-list-item-description"
-      {...props}
-    />
-  );
-}
-
-type SortableItem = {
-  id: number;
-  position: number;
-};
-
-type EditorSortableContextValue = {
-  activeId: number | null;
-  isDragging: boolean;
-};
-
-const EditorSortableContext = createContext<EditorSortableContextValue | undefined>(undefined);
-
-function EditorSortableList<T extends SortableItem>({
-  children,
-  items: initialItems,
-  onReorder,
-  renderOverlay,
-}: {
-  children: React.ReactNode;
-  items: T[];
-  onReorder: (items: { id: number; position: number }[]) => Promise<{ error: string | null }>;
-  renderOverlay?: (activeItem: T) => React.ReactNode;
-}) {
-  const dndId = useId();
-  const [mounted, setMounted] = useState(false);
-  const [activeId, setActiveId] = useState<number | null>(null);
-  const [optimisticItems, setOptimisticItems] = useOptimistic(initialItems);
-  const [pending, startTransition] = useTransition();
-
-  useEffect(() => {
-    setMounted(true);
-  }, []);
-
-  const sensors = useSensors(
-    useSensor(PointerSensor, {
-      activationConstraint: {
-        distance: 8,
-      },
-    }),
-    useSensor(TouchSensor, {
-      activationConstraint: {
-        // Short delay to distinguish tap from drag, with tolerance for slight movement
-        delay: 150,
-        tolerance: 8,
-      },
-    }),
-    useSensor(KeyboardSensor, {
-      coordinateGetter: sortableKeyboardCoordinates,
-    }),
-  );
-
-  const handleDragStart = useCallback((event: DragStartEvent) => {
-    if (typeof navigator !== "undefined" && navigator.vibrate) {
-      navigator.vibrate(10);
-    }
-    setActiveId(Number(event.active.id));
-  }, []);
-
-  const handleDragEnd = useCallback(
-    (event: DragEndEvent) => {
-      const { active, over } = event;
-      setActiveId(null);
-
-      if (!over || active.id === over.id) {
-        return;
-      }
-
-      const oldIndex = optimisticItems.findIndex((item) => item.id === active.id);
-      const newIndex = optimisticItems.findIndex((item) => item.id === over.id);
-
-      if (oldIndex === -1 || newIndex === -1) {
-        return;
-      }
-
-      const reorderedItems = arrayMove(optimisticItems, oldIndex, newIndex);
-      const newPositions = reorderedItems.map((item: SortableItem, index: number) => ({
-        id: item.id,
-        position: index,
-      }));
-
-      startTransition(async () => {
-        setOptimisticItems(reorderedItems);
-
-        const { error } = await onReorder(newPositions);
-
-        if (error) {
-          toast.error(error);
-        }
-      });
-    },
-    [optimisticItems, onReorder, setOptimisticItems],
-  );
-
-  const handleDragCancel = useCallback(() => {
-    setActiveId(null);
-  }, []);
-
-  const activeItem = activeId ? optimisticItems.find((item) => item.id === activeId) : null;
-
-  const contextValue = useMemo<EditorSortableContextValue>(
-    () => ({
-      activeId,
-      isDragging: activeId !== null,
-    }),
-    [activeId],
-  );
-
-  const itemIds = useMemo(() => optimisticItems.map((item) => item.id), [optimisticItems]);
-
-  if (!mounted) {
-    return <div data-slot="editor-sortable-list">{children}</div>;
-  }
-
-  return (
-    <EditorSortableContext.Provider value={contextValue}>
-      <DndContext
-        collisionDetection={closestCenter}
-        id={dndId}
-        onDragCancel={handleDragCancel}
-        onDragEnd={handleDragEnd}
-        onDragStart={handleDragStart}
-        sensors={sensors}
-      >
-        <SortableContext items={itemIds} strategy={verticalListSortingStrategy}>
-          <div
-            className={cn(pending && "pointer-events-none opacity-60")}
-            data-slot="editor-sortable-list"
-          >
-            {children}
-          </div>
-        </SortableContext>
-
-        <DragOverlay>
-          {activeItem && renderOverlay ? (
-            <div
-              className="bg-background rounded-md border shadow-md"
-              data-slot="editor-drag-overlay"
-            >
-              {renderOverlay(activeItem)}
-            </div>
-          ) : null}
-        </DragOverlay>
-      </DndContext>
-    </EditorSortableContext.Provider>
-  );
-}
-
-type SortableItemContextValue = {
-  attributes: ReturnType<typeof useSortable>["attributes"];
-  listeners: ReturnType<typeof useSortable>["listeners"];
-};
-
-const EditorSortableItemContext = createContext<SortableItemContextValue | undefined>(undefined);
-
-function EditorSortableItem({ children, id }: { children: React.ReactNode; id: number }) {
-  const { attributes, listeners, setNodeRef, transform, transition, isDragging } = useSortable({
-    id,
-  });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-  };
-
-  const contextValue = useMemo<SortableItemContextValue>(
-    () => ({ attributes, listeners }),
-    [attributes, listeners],
-  );
-
-  return (
-    <div
-      className={cn(isDragging && "opacity-50")}
-      data-dragging={isDragging}
-      data-slot="editor-sortable-item"
-      ref={setNodeRef}
-      style={style}
-    >
-      <EditorSortableItemContext.Provider value={contextValue}>
-        {children}
-      </EditorSortableItemContext.Provider>
-    </div>
-  );
-}
-
-function EditorSortableItemRow({ className, ...props }: React.ComponentProps<"div">) {
-  return (
-    <div
-      className={cn(
-        "group/row hover:bg-muted/50 flex items-start gap-2 px-4 py-3 transition-colors",
-        className,
-      )}
-      data-slot="editor-sortable-item-row"
-      {...props}
-    />
-  );
-}
-
-function EditorDragHandle({ children, className, ...props }: React.ComponentProps<"button">) {
-  const context = useContext(EditorSortableItemContext);
-
-  if (!context) {
-    throw new Error("EditorDragHandle must be used within an EditorSortableItem.");
-  }
-
-  return (
-    <button
-      className={cn(
-        "text-muted-foreground hover:bg-muted hover:text-foreground focus-visible:bg-muted focus-visible:text-foreground active:bg-muted relative flex min-h-11 min-w-11 shrink-0 cursor-grab touch-none items-start justify-center rounded-md pt-0.5 font-mono text-sm tabular-nums transition-colors select-none before:absolute before:top-1/2 before:left-1/2 before:size-full before:min-h-11 before:min-w-11 before:-translate-x-1/2 before:-translate-y-1/2 focus-visible:outline-none active:cursor-grabbing",
-        className,
-      )}
-      data-slot="editor-drag-handle"
-      type="button"
-      {...context.attributes}
-      {...context.listeners}
-      {...props}
-    >
-      {typeof children === "number" ? String(children).padStart(2, "0") : children}
-    </button>
-  );
-}
-
-function EditorListItemActions({
-  "aria-label": ariaLabel,
-  insertAboveLabel,
-  insertBelowLabel,
-  position,
-}: {
-  "aria-label": string;
-  insertAboveLabel: string;
-  insertBelowLabel: string;
-  position: number;
-}) {
-  const { pending, handleInsert } = useEditorList();
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger
-        aria-label={ariaLabel}
-        className="text-muted-foreground hover:bg-muted hover:text-foreground mt-0.5 flex size-6 shrink-0 items-center justify-center rounded-md opacity-0 transition-opacity group-hover/row:opacity-100 focus:opacity-100 [@media(hover:none)]:opacity-100"
-        disabled={pending}
-      >
-        <EllipsisIcon className="size-4" />
-      </DropdownMenuTrigger>
-
-      <DropdownMenuContent align="end">
-        <DropdownMenuItem onClick={() => handleInsert(position)}>
-          <PlusIcon />
-          {insertAboveLabel}
-        </DropdownMenuItem>
-
-        <DropdownMenuItem onClick={() => handleInsert(position + 1)}>
-          <PlusIcon />
-          {insertBelowLabel}
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
-
-const DEFAULT_SKELETON_COUNT = 3;
-
-function EditorListSkeleton({ count = DEFAULT_SKELETON_COUNT }: { count?: number }) {
-  return (
-    <div data-slot="editor-list">
-      <div className="flex items-center justify-end gap-2 px-4 pb-2">
-        <Skeleton className="h-8 w-28" />
-      </div>
-
-      <ul>
-        {Array.from({ length: count }).map((_, i) => (
-          <li className="border-border flex items-center gap-2 border-t px-4 py-3" key={i}>
-            <Skeleton className="size-4" />
-
-            <div className="flex min-w-0 flex-1 items-start gap-4">
-              <Skeleton className="mt-0.5 h-5 w-6" />
-
-              <div className="min-w-0 flex-1 space-y-2">
-                <Skeleton className="h-5 w-3/4" />
-                <Skeleton className="h-4 w-full" />
-              </div>
-            </div>
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-}
-
-export {
-  EditorDragHandle,
-  EditorListAddButton,
-  EditorListContent,
-  EditorListHeader,
-  EditorListItem,
-  EditorListItemActions,
-  EditorListItemContent,
-  EditorListItemDescription,
-  EditorListItemLink,
-  EditorListItemTitle,
-  EditorListProvider,
-  EditorListSkeleton,
-  EditorListSpinner,
-  EditorSortableItem,
-  EditorSortableItemRow,
-  EditorSortableList,
-};
