+++ apps/editor/src/components/editor-list/editor-list-item-title.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import { cn } from "@zoonk/ui/lib/utils";
+
+export function EditorListItemTitle({ className, ...props }: React.ComponentProps<"p">) {
+  return (
+    <p
+      className={cn("leading-tight font-medium", className)}
+      data-slot="editor-list-item-title"
+      {...props}
+    />
+  );
+}
