+++ apps/editor/src/components/editor-list/editor-sortable-context.ts
@@ -1,0 +1,17 @@
+"use client";
+
+import { createContext } from "react";
+
+export type SortableItem = {
+  id: number;
+  position: number;
+};
+
+export type EditorSortableContextValue = {
+  activeId: number | null;
+  isDragging: boolean;
+};
+
+export const EditorSortableContext = createContext<EditorSortableContextValue | undefined>(
+  undefined,
+);
