+++ apps/editor/src/components/editor-list/editor-sortable-item-context.ts
@@ -1,0 +1,13 @@
+"use client";
+
+import { type useSortable } from "@dnd-kit/sortable";
+import { createContext } from "react";
+
+export type SortableItemContextValue = {
+  attributes: ReturnType<typeof useSortable>["attributes"];
+  listeners: ReturnType<typeof useSortable>["listeners"];
+};
+
+export const EditorSortableItemContext = createContext<SortableItemContextValue | undefined>(
+  undefined,
+);
